♻️ refactor(agent-quota): move schema enums/interfaces to src/types/agentQuota
